refactor(ui): move renderSessionSummary + lifecycle into health/, barrel the tab
Finish the health.ts decomposition. renderSessionSummary moves into
health/render/session-summary.ts; loadHealthData and initHealthTab
move into health/lifecycle.ts. health.ts becomes a thin barrel
re-exporting the five public entrypoints (initHealthTab, loadHealthData,
renderHealthOverview, renderSessionSummary, renderStats) so the
app-shell wiring stays on the same import path.
